requirements.txt
autopep8==1.5.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
Pillow==7.1.2
pycodestyle==2.6.0
requests==2.23.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
